backend/requirements.txt
# T003: Backend dependencies for Todo CRUD API
# Spec Reference: plan.md Dependencies section

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database ORM
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9

# JWT Authentication
PyJWT[crypto]>=2.8.0

# Environment Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
httpx>=0.26.0
pytest-asyncio>=0.23.0
